Fix 72 security findings across all CERTIFIED S0 and S1 categories
Progress: 113/645 findings fixed (17.5%)
- All 40 CERTIFIED S0 findings now addressed
- 55 CERTIFIED S1 findings addressed (all of them)
- 72 new fixes in this commit across 40 source files

CERTIFIED S0 fixes (remaining 13 from prior commits + completions):
- Memory: CF-006 (UAF in h3_stream_lookup), CF-015 (nonce overflow),
  CF-030 (unvalidated offsets), CF-046 (slab cache for TX frames)
- Security: (all were previously fixed)
- Concurrency: CF-023 (batch-splice NAPI), CF-033 (secrets under lock),
  CF-038 (lock hierarchy), CF-045 (path ref after unlock),
  CF-047 (per-packet setkey race), CF-051 (destroy vs poll race),
  CF-055 (slab buffer size), CF-056 (sleep-in-atomic),
  CF-059 (stateless reset state machine)
- Correctness: CF-012 (stream frame ordering), CF-028 (GSO tailroom),
  CF-037 (coalesced tailroom), CF-048 (priority parsing)

CERTIFIED S1 fixes (all 55):
- Memory: CF-137, CF-148, CF-155, CF-179, CF-220, CF-226, CF-236, CF-240
- Security: CF-138, CF-139, CF-141-CF-146, CF-150-CF-151, CF-154,
  CF-156, CF-160, CF-163, CF-165-CF-166, CF-205-CF-206, CF-213
- Concurrency: CF-140, CF-147, CF-149, CF-153, CF-169, CF-174,
  CF-177-CF-178, CF-184, CF-192, CF-200, CF-202, CF-218, CF-225, CF-231
- Correctness: CF-136, CF-152, CF-157, CF-183, CF-190-CF-191, CF-195,
  CF-211, CF-217, CF-233, CF-237
- Perf: CF-180

Key changes by subsystem:
- crypto/: Secret zeroization, AEAD pre-allocation, bounds checks,
  key update race fixes, ASN.1 parser hardening
- core/: Packet parsing bounds, stream state machine, ECN overflow,
  priority parsing, connection accept heap alloc
- http3/: Stream refcounting, settings varint encoding, TOCTOU fixes,
  QPACK overflow/memory limits, WebTransport safe iteration
- input/output: GRO CID length fix, batch-splice NAPI, slab cache
  for TX frames, GSO/coalesce tailroom checks, stateless reset
- security: Constant-time CID, timing normalization, capability checks
- misc: Lock hierarchy standardization, sleep-in-atomic, lockless
  io_uring ring, zerocopy retry limits, per-CPU GRO stats

+++ net/tquic/quic_offload.c
@@ -163,9 +163,12 @@ static int tquic_parse_header(struct sk_buff *skb, struct tquic_offload_cb *cb)
 		if (skb->len < 6)
 			return -EINVAL;
 
-		/* Version field */
-		version = (data[1] << 24) | (data[2] << 16) |
-			  (data[3] << 8) | data[4];
+		/* Version field - cast to u32 before shifting to avoid
+		 * undefined behavior when u8 is promoted to int and
+		 * shifted left by 24 (sign bit overflow).
+		 */
+		version = ((u32)data[1] << 24) | ((u32)data[2] << 16) |
+			  ((u32)data[3] << 8) | (u32)data[4];
 		offset = 5;
 
 		/* DCID length */
